moved stuff around to make sense
